pyproject.toml
[tool.poetry]
name = "PSFL"
version = "0.1.0"
description = ""
authors = ["Davide Domini <davide.domini@unibo.it>"]
readme = "README.md"

packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.13"
phyelds = "2.6.0"
torch = "^2.7.0"
torchvision = "^0.22.0"
profed = "0.6.5"
pyyaml = "^6.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
